// libs/features/contributor/src/lib/shared/enums/contributor-banner-fields.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum ContributorBannerFields {
  NAME_BILINGUAL = 'nameBilingual',
  NAME_ENGLISH = 'nameEnglish',
  NAME_ARABIC = 'nameArabic',
  IDENTIFIER = 'identifier',
  DOB = 'dob',
  MARITAL_STATUS = 'maritalStatus',
  GENDER = 'gender',
  NATIONALITY = 'nationality',
  STATUS = 'status'
}